requirements.txt
langchain>=0.1.0
langchain-google-genai>=1.0.0
langchain-pinecone>=0.1.0
langchain-community>=0.0.20
pinecone-client>=3.0.0
pypdf>=3.17.0
python-dotenv>=1.0.0
tqdm>=4.66.0